+++ app/javascript/lca/utils/calculated/_pools.js
@@ -70,6 +70,9 @@ export function disengage(character, merits, penalties, excellencyAbils) {
   let bonus = []
   if (merits.some((m) => m.startsWith('fleet of foot')))
     bonus = [{ label: 'fleet of foot', bonus: 1 }]
+  if (character.type !== 'Character' && character.caste === 'water')
+    bonus = bonus.concat([{ label: '/3m anima', bonus: 1, situational: true }])
+
   const penalty = [{ label: 'wound', penalty: penalties.wound }, { label: 'mobility', penalty: penalties.mobility }]
   return pool('Disengage', character, 'dexterity', 'dodge', bonus, penalty, excellencyAbils)
 }